Recombine all similar tests into a single module
